requirements.txt
functions-framework==3.*
google-cloud-storage
requests
python-dotenv